Numerical.BLAS.Single

Description

This module provides BLAS level-1 library functions for vectors of single precision floating point numbers. In general, BLAS level-1 functions that take vector valued arguments support the ability to apply their functionality over all the elements in the vector, or to a smaller vector comprised of elements drawn in uniform intervals from the original vector argument. This capability allows vector operations to be applied, for example, to columns of matrices.

To support this capability, each BLAS level-1 function that has at least one vector valued argument takes an Int parameter, n, that represents the number of elements to be extracted from the vector arguments. The function also will take an additional Int argument for each vector valued argument, representing the both the direction of traversal and the interval between elements are drawn of the original vector argument. Typically n is the first Int valued argument, and the increments are the first Int valued argment following a vector agument.

For example:

sdot :: Int -> Vector Float -> Int -> Vector Float -> Int -> Float
sdot n sx incx sy incy = foldl' (+) 0 $ zipWith (*) (sampleElems n sx incx) (sampleElems n sy incy)

is morally equivalent to the dot product of the vectors sx and sy when both vectors have lengths of n and both incx and incy are unity.

The BLAS functions do not generally guard against invalid values of the increment arguments. For each vector argument v, and each increment inc, the calling program should ensure that:

inc /= 0
length v >= (1 + (n-1)*inc)

Norms

asum

Arguments

:: (Num a, Storable a)

=> Int
The number of elements to sum, n

-> Vector a
The vector from which summands are selected, x

-> Int
The increment between summands, incx

-> a

O(n) - asum computes the sum of the absolute value of elements drawn a vector, x. This function is morally equivalent to the L1 norm when n is equal to the length of x and incx is unity.

sasum :: Int -> Vector Float -> Int -> Float

Deprecated: Use asum instead

dasum :: Int -> Vector Double -> Int -> Double

Deprecated: Use asum instead

nrm2

Arguments

:: (Floating a, Ord a, Storable a)

=> Int
The number of elements to sum, n

-> Vector a
The vector from which summands are selected, x

-> Int
The increment between summands, incx

-> a

O(n) - nrm2 computes the sum of the squared values of elements drawn a vector, x. This function is morally equivalent to the L2 norm when n is equal to the length of x and incx is unity.

snrm2 :: Int -> Vector Float -> Int -> Float

Deprecated: Use nrm2 instead

dnrm2 :: Int -> Vector Double -> Int -> Double

Deprecated: Use nrm2 instead

Products

dot

Arguments

:: (Storable a, Num a)

=> Int
The number of elements n

-> Vector a
The first vector, x

-> Int
The increment, incx, between elements drawn from x

-> Vector a
The second vector, y

-> Int
The increment, incy, between elements drawn from y

-> a
The sum of products of the selected elements

O(n) - compute the sum of products of elements drawn from two Float vectors. This function is morally equivalent to the dot product of vectors x and y when n is the length of both x and y, and both incx and incy are unity.

sdot :: Int -> Vector Float -> Int -> Vector Float -> Int -> Float

Deprecated: Use dot instead.

ddot :: Int -> Vector Double -> Int -> Vector Double -> Int -> Double

Deprecated: Use dot instead.

sdsdot

Arguments

:: Int
The number of elements n

-> Float
a Vector valued adjustment to add to the sum

-> Vector Float
The first vector, x

-> Int
The increment, incx, between elements drawn from x

-> Vector Float
The second vector, y

-> Int
The increment, incy, between elements drawn from y

-> Float
The sum of products of the selected elements

O(n) - compute the sum of products of elements drawn from two Float vectors using Double precision internally. This function is morally equivalent to the dot product of vectors x and y when n is the length of both x and y, and both incx and incy are unity.

scal

Arguments

:: (Storable a, Num a)

=> Int
The number of elements to scale, n

-> a
The scalar factor, a

-> Vector a
The vector to scale, x

-> Int
The increment between elements to scale, incx

-> Vector a
The vector with scaled elements

O(n) - Multiply selected elements in a vector, x, by a scalar, a. This function is morally equivalent to scalar multiplication a*x when n is the length x and incx is unity.

sscal :: Int -> Float -> Vector Float -> Int -> Vector Float

Deprecated: Use scal instead.

dscal :: Int -> Double -> Vector Double -> Int -> Vector Double

Deprecated: Use scal instead.

Rotations

rot

Arguments

:: (Storable a, Num a)

=> Int
The number of elements to rotate

-> Vector a
The first vector to rotate, x

-> Int
The interval between rotated elements in the first vector, incx

-> Vector a
The second vector to rotate, y

-> Int
The interval between rotated elements in the second vector, incy

-> a
The cosine of the angle of rotation

-> a
The sine of the angle of rotation

-> (Vector a, Vector a)

O(n) apply a plane rotation to elements drawn from a pair of vectors. This is morally equivalent to rotating both vectors through an angle when n is equal to the length of both the vectors, and both incx and incy are unity.

srot :: Int -> Vector Float -> Int -> Vector Float -> Int -> Float -> Float -> (Vector Float, Vector Float)

Deprecated: Use rot instead.

drot :: Int -> Vector Double -> Int -> Vector Double -> Int -> Double -> Double -> (Vector Double, Vector Double)

Deprecated: Use rot instead.

rotg

Arguments

:: (Fractional a, Floating a, Ord a)

=> a
The x coordinate of the deconstructed vector

-> a
The y coordinate of the deconstructed vector

-> GivensRot a
The Givens rotation coefficients

O(1) Construct a plane Givens rotation on a deconstructed two-vector. Specifically:

rotg sa sb = ( r, secant(theta), cos(theta), sin(theta))

where r is the signed magnitude of the vector sa,sb. In the case when r is zero, rotg returns (0,0,1,0). See also "rot".

srotg :: Float -> Float -> GivensRot Float

Deprecated: Use rotg instead.

drotg :: Double -> Double -> GivensRot Double

Deprecated: Use rotg instead.

rotm :: (Storable a, Num a) => ModGivensRot a -> Int -> Vector a -> Int -> Vector a -> Int -> (Vector a, Vector a)

O(n) apply a modified Givens rotation to elements drawn from a pair of vectors according to the following specification:

srotm FLAGNEG2 n sx incx sy incy = (sx,sy)
srotm (FLAGNEG1 h11 h12 h21 h22) n sx incx sy incy =
   ( updateElems ( x y -> h11*x + h12*y) n sx incx sy incy
   , updateElems ( y x -> h21*x + h22*y) n sy incy sx incx)
srotm (FLAG0 h12 h21) n sx incx sy incy =
   ( updateElems ( x y -> x + h12*y) n sx incx sy incy
   , updateElems ( y x -> h21*x + y) n sy incy sx incx)
srotm (FLAG1 h11 h22) n sx incx sy incy =
   ( updateElems ( x y -> h11*x + y) n sx incx sy incy
   , updateElems ( y x -> -x + h22*y) n sy incy sx incx)

see also "rotmg".

srotm :: ModGivensRot Float -> Int -> Vector Float -> Int -> Vector Float -> Int -> (Vector Float, Vector Float)

Deprecated: Use rotm instead

drotm :: ModGivensRot Double -> Int -> Vector Double -> Int -> Vector Double -> Int -> (Vector Double, Vector Double)

Deprecated: Use rotm instead

rotmg

Arguments

:: (Ord a, Fractional a)

=> a
The first diagonal element sd1

-> a
The second diagonal element sd2

-> a
The first component of the vector sx1

-> a
The second component of the vector sx2

-> ModGivensRot a
The modified Givens Rotation coefficients

O(1) Construct a modified plane Givens rotation on a deconstructed two-vector as described in "Basic Linear Algebra Subprograms for Fortran Use", Lawson 1979. The solution should satisfy

| d1 0 | | h11 h12 | | sx1/x1 | | sd1 0 | | sx1 | | 0 d2 | | h21 h22 | | sy1/x1 | = G | 0 sd2 | | sx2 | @

where G is the Givens rotation matirx. d1, d2 and x1 are rescaled by the the "checkscale" subprogram to be within the conservative limits of +- 1(4096)^2 and +/- 4096^2.

See also "rotm".

srotmg :: Float -> Float -> Float -> Float -> ModGivensRot Float

Deprecated: Use rotmg instead.

drotmg :: Double -> Double -> Double -> Double -> ModGivensRot Double

Deprecated: Use rotmg instead.

Linear Combinations

axpy

Arguments

:: (Storable a, Num a)

=> Int
The number of elements

-> a
The scalar multiple for the first vector, a

-> Vector a
The first vector, x

-> Int
The interval between elements drawn from the first vector, incx

-> Vector a
The second vector, y

-> Int
The interval between elements drawn from the second vector, incy

-> Vector a
The linear combination a*x + y

O(n) compute the linear combination of elements drawn from two vectors according. This is morally equivalent to the linear combination of the two vectors when n is the length of the vectors and both incx and incy are unity.

saxpy :: Int -> Float -> Vector Float -> Int -> Vector Float -> Int -> Vector Float

Deprecated: Use axpy instead.

daxpy :: Int -> Double -> Vector Double -> Int -> Vector Double -> Int -> Vector Double

Deprecated: Use axpy instead.

Memory Operations

copy

Arguments

:: Storable a

=> Int
The number of elements to copy

-> Vector a
The source vector from which to copy elements

-> Int
The increment between elements drawn from the source vector

-> Vector a
The target vector into which elements are copied

-> Int
The increment between the destination elements

-> Vector a

O(n) - Copy n elements from one Float vector into another.

scopy :: Int -> Vector Float -> Int -> Vector Float -> Int -> Vector Float

Deprecated: Use copy instead

dcopy :: Int -> Vector Double -> Int -> Vector Double -> Int -> Vector Double

Deprecated: Use copy instead

swap

Arguments

:: Storable a

=> Int
The number of elements to copy

-> Vector a
The first vector from which elements will be swapped

-> Int
The increment between elements drawn from the first vector

-> Vector a
The second vector from which elements will be swapped

-> Int
The increment between elements drawn from the second vector

-> (Vector a, Vector a)

O(n) -- Swap n elements between two Float vectors.

sswap :: Int -> Vector Float -> Int -> Vector Float -> Int -> (Vector Float, Vector Float)

Deprecated: Use swap instead.

dswap :: Int -> Vector Double -> Int -> Vector Double -> Int -> (Vector Double, Vector Double)

Deprecated: Use swap instead.

Specialized Folds

iamax

Arguments

:: (Ord a, Num a, Storable a)

=> Int
The number of elements drawn from the source vector,n

-> Vector a
The source vector, x

-> Int
The interval between elements drawn from the source vector, incx

-> Int
The index of the element with the largest absolute value.

O(n) - Find the index of the element with the largest absolute value from elements drawn from a vector. This function is morally equivalent to the index of the element with the largest magnitude when n is the length of the vector x and the increment, inc, is unity.

isamax :: Int -> Vector Float -> Int -> Int

Deprecated: use iamax instead

idamax :: Int -> Vector Double -> Int -> Int

Deprecated: use iamax instead