Numerical.BLAS.Types

Documentation

data GivensRot a

Constructors

GIVENSROT !(a, a, a, a)

data ModGivensRot a

ModGivensRot is a data structure for containing the linear operator obtained from the rotmg function. The values h11, h12, h21, and h22 are the elements of the linear operator. The four cases correspond two the value of the SFLAG.

Constructors

FLAGNEG2

FLAGNEG1

Fields

d1, d2, x1, h11, h12, h21, h22 :: !a

FLAG0

Fields

d1, d2, x1, h12, h21 :: !a

FLAG1

Fields

d1, d2, x1, h11, h22 :: !a